requirements.txt
mysql-connector
SQLAlchemy~=1.3.23
paho-mqtt